layer files separation, maximize and minimize layers functionality
