+++ src/iartisanxl/buttons/expand_contract_button.py
@@ -1,19 +1,23 @@
 from importlib.resources import files
+
 from PyQt6 import QtGui
-from PyQt6.QtCore import Qt, QRectF
+from PyQt6.QtCore import QRectF, Qt
 
 from iartisanxl.buttons.base_menu_button import BaseMenuButton
 
 
-class ExpandRightButton(BaseMenuButton):
+class ExpandContractButton(BaseMenuButton):
     CONTRACT_ICON = files("iartisanxl.theme.icons").joinpath("chevron_right.png")
     EXPAND_ICON = files("iartisanxl.theme.icons").joinpath("chevron_left.png")
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
+    def __init__(self, button_width: int = 40, button_height: int = 40, extended: bool = True, inverted: bool = False):
+        super().__init__()
 
-        self.extended = True
-        self.setFixedSize(40, 44)
+        self.button_width = button_width
+        self.button_height = button_height
+        self.inverted = inverted
+        self.extended = extended
+        self.setFixedSize(self.button_width, self.button_height + 4)
 
     def paintEvent(self, _event):
         painter = QtGui.QPainter(self)
@@ -24,12 +28,18 @@ class ExpandRightButton(BaseMenuButton):
         )
 
         if self.extended:
-            icon = QtGui.QIcon(str(self.CONTRACT_ICON))
+            if self.inverted:
+                icon = QtGui.QIcon(str(self.CONTRACT_ICON))
+            else:
+                icon = QtGui.QIcon(str(self.EXPAND_ICON))
         else:
-            icon = QtGui.QIcon(str(self.EXPAND_ICON))
+            if self.inverted:
+                icon = QtGui.QIcon(str(self.EXPAND_ICON))
+            else:
+                icon = QtGui.QIcon(str(self.CONTRACT_ICON))
         icon.paint(
             painter,
-            QRectF(0, 2, 40, 40).toRect(),
+            QRectF(0, 2, self.button_width, self.button_height).toRect(),
             Qt.AlignmentFlag.AlignCenter,
             state=QtGui.QIcon.State.Off,
         )
